Remove initAll() from TaskPage and add to TaskDetailsPage
Whilst the details component was closing in TaskPage it wasn't opening when clicking the link so it's been moved to TaskDetailsPage which was working in both cases.

Ideally in the future we need to look to see if there is a better place to put initAll rather than putting it in several places.