Get the extended staff details to save in the session
Add an http request to get the extended staff details, which will return the line manager and delegate emails. This is called when the user starts/edits a shift and the data is saved in local storage with a key of extendedStaffDetailsContext.

Also updated the unit tests.

This has been tested locally with the User Profile and Man Dec Make Declaration forms