Add browser detection
The idea here is that we can, if we like, specify browsers for which a minimum version is required. We expect to make such a specification only for Chrome and Edge at the moment.

In this commit, the on/off flag and minimum versions are hard-coded, but this is only temporary. In another PR soon these will be driven by environment variables.

Also in this commit some minor tidy-up and accommodation for optional Prettier formatting.