Fetch staffid for shift form
This will no longer be retrieved from the staff details database query nor from the extended staff details database query, yet it the shift form expects it to be available. Therefore, for now at least, we retrieve it from the operational data matching against the user's email from Keycloak and merge it with staffDetails both in the store and in local storage.